Change ServiceAdvertiser and move PeersProvider
Remove the ServiceAdvertiser.onPeersChanged method which was never used.
Moved the PeersProvider interface back to the servicediscovery.webrtc package.
