+++ babble/src/androidTest/java/io/mosaicnetworks/babble/discovery/HttpPeerDiscoveryTest.java
@@ -7,7 +7,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 import io.mosaicnetworks.babble.node.Peer;
 import io.mosaicnetworks.babble.servicediscovery.mdns.HttpPeerDiscoveryRequest;
 import io.mosaicnetworks.babble.servicediscovery.mdns.HttpPeerDiscoveryServer;
-import io.mosaicnetworks.babble.node.PeersProvider;
+import io.mosaicnetworks.babble.servicediscovery.mdns.PeersProvider;
 import io.mosaicnetworks.babble.servicediscovery.mdns.ResponseListener;
 
 import org.junit.Rule;
