+++ babble/src/main/java/io/mosaicnetworks/babble/node/BabbleNode.java
@@ -30,6 +30,7 @@ import com.google.gson.JsonSyntaxException;
 
 import java.nio.charset.StandardCharsets;
 
+import io.mosaicnetworks.babble.servicediscovery.mdns.PeersProvider;
 import mobile.Mobile;
 import mobile.Node;
